nibble functions needed uint parameter or things get sign extended with -O
git-svn-id: https://vgmstream.svn.sourceforge.net/svnroot/vgmstream@6 51a99a44-fe44-0410-b1ba-c3e57ba2b86b

+++ src/util.h
@@ -28,12 +28,12 @@ static inline int32_t get_32bitLE(uint8_t * p) {
 /* signed nibbles come up a lot */
 static int nibble_to_int[16] = {0,1,2,3,4,5,6,7,-8,-7,-6,-5,-4,-3,-2,-1};
 
-static inline int get_high_nibble_signed(int n) {
+static inline int get_high_nibble_signed(uint8_t n) {
     /*return ((n&0x70)-(n&0x80))>>4;*/
     return nibble_to_int[n>>4];
 }
 
-static inline int get_low_nibble_signed(int n) {
+static inline int get_low_nibble_signed(uint8_t n) {
     /*return (n&7)-(n&8);*/
     return nibble_to_int[n&0xf];
 }
